requirements.txt
absl-py==2.3.1
altair==4.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
entrypoints==0.4
exceptiongroup==1.3.1
facenet-pytorch==2.6.0
fastapi==0.128.3
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-pasta==0.2.0
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
libclang==18.1.1
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.4
moviepy==2.2.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.4.2
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
pillow==10.2.0
proglog==0.1.12
protobuf==4.25.8
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.1
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sounddevice==0.5.5
starlette==0.50.0
streamlit==1.22.0
sympy==1.13.1
tenacity==8.5.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.4
tqdm==4.67.1
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==2.0.1
zipp==3.23.0
